Improve handling of text size preference
Preference changes are now handled via a callback in the GreekTextView
class. The text size can now be fixed at medium by setting the
allowTextSizeChanges attribute to false in GreekTextView XML layouts.
This functionality is used to prevent the text size from changing in
ListView items.
