+++ GreekReference/src/main/java/com/benlinskey/greekreference/AbstractDetailFragment.java
@@ -31,46 +31,7 @@ public abstract class AbstractDetailFragment extends Fragment {
     
     private static final String KEY_SCROLL_Y = "scroll_y";
 
-    /**
-     * An enum type used to store text size values.
-     * <p>
-     * Note that the scaled pixel values stored here correspond to the standard Android
-     * TextAppearance values defined in the SDK data/res/values/styles.xml file.
-     */
-    private enum TextSize {
-        SMALL("Small", 14),
-        MEDIUM("Medium", 18),
-        LARGE("Large", 22);
-
-        private final String mName; // Size name stored in preference array
-        private final float mSize;  // Text size in scaled pixels
-
-        /**
-         * Enum type constructor.
-         * @param name the name of the size defined in the preferences array
-         * @param size the text size in scaled pixels
-         */
-        private TextSize(String name, float size) {
-            mName = name;
-            mSize = size;
-        }
-
-        /**
-         * Returns the text size in scaled pixels for the specified size.
-         * @param name the name of the size defined in the preferences array
-         * @return the corresponding text size in scaled pixels
-         */
-        public static float getScaledPixelSize(String name) {
-            for (TextSize tx : TextSize.values()) {
-                if (name.equals(tx.mName)) {
-                    return tx.mSize;
-                }
-            }
-            throw new IllegalArgumentException("Invalid text size name.");
-        }
-    }
-
-    // We use a single Toast object to prevent overlapping toasts when the user 
+    // We use a single Toast object to prevent overlapping toasts when the user
     // repeatedly taps an icon that displays a toast.
     protected Toast mToast;
 
@@ -87,28 +48,7 @@ public abstract class AbstractDetailFragment extends Fragment {
             int scrollY = savedInstanceState.getInt(KEY_SCROLL_Y);
             View scrollView = getActivity().findViewById(R.id.detail_scroll_view);
             scrollView.setScrollY(scrollY);
-            setTextSize();
         }
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        setTextSize();
-    }
-
-    // TODO: Move this method to the GreekTextView class.
-    /**
-     * Sets the {@code GreekTextView}'s text size to the size stored in the preferences.
-     */
-    private void setTextSize() {
-        SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(getActivity());
-        String textSizeKey = getString(R.string.pref_textSize_key);
-        String defaultSize = getString(R.string.pref_textSize_item_medium);
-        String textSize = prefs.getString(textSizeKey, defaultSize);
-        float sp = TextSize.getScaledPixelSize(textSize);
-        GreekTextView textView = (GreekTextView) getActivity().findViewById(R.id.item_detail);
-        textView.setTextSize(sp);
     }
 
     @Override
